requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.1.3
django-crispy-forms==1.14.0
django-environ==0.9.0
django-ranged-response==0.2.0
django-smart-selects==1.6.0
django-taggit==3.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
sqlparse==0.4.3

